Use float instead of int for heat values

+++ satviz-consumer/src/main/java/edu/kit/satviz/consumer/graph/HeatUpdate.java
@@ -2,7 +2,7 @@ package edu.kit.satviz.consumer.graph;
 
 public class HeatUpdate implements GraphUpdate {
 
-  public void add(int index, int heat) {
+  public void add(int index, float heat) {
 
   }
